+++ prboom2/src/v_video.h
@@ -378,7 +378,7 @@ int V_FillHeightVPT(int y, int height, byte color, enum patch_translation_e flag
 // heretic
 
 void V_DrawRawScreen(const char *lump_name);
-void V_DrawRawScreenOffset(const char *lump_name, int x_offset, int y_offset, enum patch_translation_e flags);
+void V_DrawRawScreenOffset(const char *lump_name, float x_offset, float y_offset, enum patch_translation_e flags);
 void V_DrawShadowedNumPatch(int x, int y, int lump);
 void V_DrawShadowedNamePatch(int x, int y, const char* name);
 void V_DrawTLNumPatch(int x, int y, int lump);
